requirements.txt
Django==1.10
-e git://github.com/dwaiter/django-bcrypt.git@913d86b2ba71334fd54670c6f0050bec02689654#egg=django_bcrypt
django-extensions==1.8.1
py-bcrypt==0.4
pyparsing==2.2.0
six==1.10.0
